Update events to emit model/session objects
This commit updates the events emitted in cli driver
to emit service/operation object alternatives.  This change
allows handlers to be updated to not require service/operation
objects.

In many cases, some handlers took an `operation` object, but
did not actually use this, so these changes were minimal.

Note that this puts the CLI in an interim state.  The plan is
to not merge this into the develop branch until clidriver is
switched over, but this will allow a common base for all
customization switchovers to be based on.